# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Actual Bridge API
  version: 1.0.0
  description: |
    Stable contract between the Android app and the Actual bridge.

    Conventions
    - All money values are integers in minor units (cents). Ledger amounts are signed:
      outflows negative, inflows positive. Budget amounts (`budgeted`, `totalBudgeted`)
      are the amount assigned, positive when money is assigned.
    - Months are `YYYY-MM`, dates are `YYYY-MM-DD`.
    - Within /v1 changes are additive only. Clients must ignore unknown fields and
      treat unknown enum values as `unknown`.
    - Non-idempotent writes require an `Idempotency-Key` header. Creates use a
      client-generated UUID `id` instead.
    - Errors are `application/problem+json` with a stable `code`.
servers:
  - url: https://budget-api.example.com
security:
  - bearer: []
...
paths:
...
  /v1/me/home:
    get:
      tags: [household]
      operationId: getHomeLayout
      summary: This person's Home layout (widget ids in order, and hidden ones). Empty lists mean the default.
      responses:
        '200':
          description: Layout
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeLayout' }
    put:
      tags: [household]
      operationId: setHomeLayout
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HomeLayout' }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeLayout' }
  /v1/budgets/{budgetId}/appearance:
    get:
      tags: [budgets]
      operationId: getAppearance
      summary: Category colors and emoji chosen by the household (Actual has none). Categories not listed use the app's defaults.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '200':
          description: Appearance
          content:
            application/json:
              schema:
                type: object
                required: [categories]
                properties:
                  categories: { type: array, items: { $ref: '#/components/schemas/CategoryAppearance' } }
  /v1/budgets/{budgetId}/categories/{categoryId}/appearance:
    put:
      tags: [budgets]
      operationId: setCategoryAppearance
      summary: Members. A category or a category group. Either can be set alone; a missing or null one means the default. Both null (or {}) goes back to the defaults.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: categoryId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                color: { type: [string, 'null'], pattern: '^#[0-9A-Fa-f]{6}$' }
                emoji: { type: [string, 'null'] }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryAppearance' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    BudgetId:
      name: budgetId
      in: path
      required: true
      description: Actual sync id (the budget's `groupId`), stable across devices
      schema: { type: string }
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Stable machine-readable code, e.g. `unauthorized`, `forbidden`, `not_found`, `validation`, `feature_unavailable`, `actual_unavailable`, `budget_encrypted`, `bank_sync_failed`, `update_failed`, `conflict`, `rate_limited`
        detail: { type: string }
...
    HomeLayout:
      type: object
      required: [order, hidden]
      properties:
        order: { type: array, items: { type: string }, description: Widget ids in the order chosen; widgets not listed follow in their default order }
        hidden: { type: array, items: { type: string } }
    CategoryAppearance:
      type: object
      required: [categoryId, color, emoji]
      properties:
        categoryId: { type: string }
        color: { type: [string, 'null'], description: '#RRGGBB' }
        emoji: { type: [string, 'null'] }
...